# docs/swagger.yaml
basePath: /
definitions:
...
  dto.BlogResponse:
    properties:
      author:
        $ref: '#/definitions/dto.UserBriefResponse'
      authorId:
        type: string
      category:
        $ref: '#/definitions/dto.CategoryResponse'
      categoryId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      excerpt:
        type: string
      id:
        type: string
      publishedAt:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/entity.BlogStatus'
      tags:
        items:
          $ref: '#/definitions/dto.TagResponse'
        type: array
      thumbnailUrl:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      visibility:
        $ref: '#/definitions/entity.BlogVisibility'
    type: object
  dto.CategoryResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.CreateBlogRequest:
    properties:
      categoryId:
        type: string
      content:
        type: string
      excerpt:
        type: string
      slug:
        maxLength: 255
        minLength: 1
        type: string
      tagIds:
        items:
          type: string
        type: array
      thumbnailUrl:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - content
    - slug
    - title
    type: object
...
  dto.PublishBlogRequest:
    properties:
      visibility:
        enum:
        - public
        - subscribers_only
        type: string
    required:
    - visibility
    type: object
...
  dto.TagResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  dto.UpdateBlogRequest:
    properties:
      categoryId:
        type: string
      content:
        type: string
      excerpt:
        type: string
      slug:
        maxLength: 255
        minLength: 1
        type: string
      tagIds:
        items:
          type: string
        type: array
      thumbnailUrl:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    type: object
...
  dto.UserBriefResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  entity.BlogStatus:
    enum:
    - draft
    - published
    type: string
    x-enum-varnames:
    - BlogStatusDraft
    - BlogStatusPublished
  entity.BlogVisibility:
    enum:
    - public
    - subscribers_only
    type: string
    x-enum-varnames:
    - BlogVisibilityPublic
    - BlogVisibilitySubscribersOnly
  response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: Clean Architecture Go API with GORM, Gin, PostgreSQL, Redis
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /api/v1/blogs:
    get:
      consumes:
      - application/json
      description: List blogs with optional filters
      parameters:
      - description: Filter by author ID
        in: query
        name: authorId
        type: string
      - description: Filter by category ID
        in: query
        name: categoryId
        type: string
      - description: Filter by status (draft, published)
        in: query
        name: status
        type: string
      - description: Filter by visibility (public, subscribers_only)
        in: query
        name: visibility
        type: string
      - description: Search in title or content
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: List blogs
      tags:
      - Blogs
    post:
      consumes:
      - application/json
      description: Create a new blog post (draft by default)
      parameters:
      - description: Blog data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBlogRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BlogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a new blog
      tags:
      - Blogs
...
  /api/v1/blogs/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a blog (author only)
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a blog
      tags:
      - Blogs
    get:
      consumes:
      - application/json
      description: Get a blog post by its ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get blog by ID
      tags:
      - Blogs
    put:
      consumes:
      - application/json
      description: Update an existing blog (author only)
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      - description: Blog data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBlogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update a blog
      tags:
      - Blogs
  /api/v1/blogs/{id}/publish:
    post:
      consumes:
      - application/json
      description: Publish a draft blog with visibility setting
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      - description: Visibility setting
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PublishBlogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Publish a blog
      tags:
      - Blogs
  /api/v1/blogs/{id}/unpublish:
    post:
      consumes:
      - application/json
      description: Revert a published blog back to draft
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BlogResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Unpublish a blog
      tags:
      - Blogs
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"